+++ packages/dashboard/src/components/layout/drawer/drawerHeader.tsx
@@ -1,6 +1,5 @@
 // material-ui
 import { Stack } from "@mui/material";
-import { useTheme } from "@mui/material/styles";
 
 import Profile from "../header/headerContent/profile";
 // project import
@@ -9,10 +8,8 @@ import DrawerHeaderStyled from "./drawerHeader/drawerHeaderStyled";
 // ==============================|| DRAWER HEADER ||============================== //
 
 function DrawerHeader({ open }: { open: boolean }) {
-  const theme = useTheme();
-
   return (
-    <DrawerHeaderStyled theme={theme} open={open}>
+    <DrawerHeaderStyled open={open}>
       <Stack direction="row" spacing={1} alignItems="center">
         <Profile />
       </Stack>
