--- packages/dashboard/src/components/layout/drawer/drawerHeader/drawerHeaderStyled.ts
@@ -1,17 +1,0 @@
-// material-ui
-import { Box, Theme } from "@mui/material";
-import { styled } from "@mui/material/styles";
-
-// ==============================|| DRAWER HEADER - STYLED ||============================== //
-
-const DrawerHeaderStyled = styled(Box, {
-  shouldForwardProp: (prop) => prop !== "open",
-})(({ theme, open }: { theme: Theme; open: boolean }) => ({
-  ...theme.mixins.toolbar,
-  display: "flex",
-  alignItems: "center",
-  justifyContent: open ? "flex-start" : "center",
-  paddingLeft: theme.spacing(open ? 3 : 0),
-}));
-
-export default DrawerHeaderStyled;
