+++ packages/dashboard/src/components/layout/drawer/drawerHeader/drawerHeaderStyled.tsx
@@ -1,0 +1,30 @@
+// material-ui
+import { Box, useTheme } from "@mui/material";
+import { ComponentProps } from "react";
+
+// ==============================|| DRAWER HEADER - STYLED ||============================== //
+
+interface DrawerComponentProps {
+  open: boolean;
+}
+
+function DrawerHeaderStyled({
+  open,
+  ...props
+}: DrawerComponentProps & ComponentProps<typeof Box>) {
+  const theme = useTheme();
+  return (
+    <Box
+      sx={{
+        ...theme.mixins.toolbar,
+        display: "flex",
+        alignItems: "center",
+        justifyContent: open ? "flex-start" : "center",
+        paddingLeft: theme.spacing(open ? 3 : 0),
+      }}
+      {...props}
+    />
+  );
+}
+
+export default DrawerHeaderStyled;
